add launch file support, convert GameEngine to package, move all simulation files to one place
