+++ requirements.txt
@@ -3,3 +3,4 @@ numpy
 pandas
 matplotlib
 pyyaml
+tqdm
